+++ src/index.ts
@@ -1,2 +1,2 @@
 export { useFetch, useRefetch } from './useFetch';
-export { FetchProvider } from './context';
+export { FetchProvider } from './FetchProvider';
