--- src/types.ts
@@ -1,8 +1,0 @@
-export type FetchFunc<Input, Result> = (input: Input) => Promise<Result>;
-
-export type FetchState<Input, Result> = {
-  input: Input;
-  result?: Result;
-  error?: unknown;
-  promise?: Promise<void>;
-}
